+++ crates/zng-wgt-window/src/window_properties.rs
@@ -462,3 +462,24 @@ pub fn config_block_window_load(child: impl UiNode, enabled: impl IntoValue<Bloc
         _ => {}
     })
 }
+
+context_var! {
+    /// Variable that indicates the context should use mobile UI themes.
+    ///
+    /// This is `true` by default in Android and iOS builds. It can also be set using [`force_mobile`](fn@force_mobile).
+    pub static IS_MOBILE_VAR: bool = cfg!(any(target_os = "android", target_os = "ios"));
+}
+
+/// Requests mobile UI themes in desktop builds.
+///
+/// This property sets the [`IS_MOBILE_VAR`].
+#[property(CONTEXT, default(IS_MOBILE_VAR), widget_impl(Window))]
+pub fn force_mobile(child: impl UiNode, is_mobile: impl IntoVar<bool>) -> impl UiNode {
+    with_context_var(child, IS_MOBILE_VAR, is_mobile)
+}
+
+/// Gets the [`IS_MOBILE_VAR`] that indicates the window or widget should use mobile UI themes.
+#[property(EVENT, widget_impl(Window))]
+pub fn is_mobile(child: impl UiNode, is_mobile: impl IntoVar<bool>) -> impl UiNode {
+    zng_wgt::node::bind_state(child, IS_MOBILE_VAR, is_mobile)
+}