+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Add `IS_MOBILE_VAR`, `is_mobile` and `force_mobile` context var and properties.
 * Add `WindowVars::safe_padding` and implement it for Android.
 * **Breaking** Add `WindowOpenData::safe_padding` and `WindowChanged::safe_padding` to the view API.
     - Note that this is only a breaking change for direct dependents of `zng-view-api`.